+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * [UIBULKED-377](https://issues.folio.org/browse/UIBULKED-377) Not all note types are displayed using "Change note type".
 * [UIBULKED-372](https://issues.folio.org/browse/UIBULKED-372) Bulk edit - Element IDs are not unique.
 * [UIBULKED-371](https://issues.folio.org/browse/UIBULKED-371) Bulk edit - Grouped form controls missing accessible name.
+* [UIBULKED-353](https://issues.folio.org/browse/UIBULKED-353) Update Electronic access - URI.
 
 ## [4.0.0](https://github.com/folio-org/ui-bulk-edit/tree/v4.0.0) (2023-10-12)